+++ Source/Core/Log/NovusLogger.Console.pas
@@ -1,0 +1,16 @@
+unit NovusLogger.Console;
+
+interface
+
+Uses NovusLogger, NovusLog;
+
+type
+  TNovusLogger_Target_Console = class(TNovusLogger_Target)
+  private
+  protected
+  Public
+  end;
+
+implementation
+
+end.